backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.0
anthropic==0.7.0
# qwen-api - Install separately if needed: pip install dashscope
transformers==4.35.0
torch>=2.0.0
datasets==2.14.6
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0
python-multipart==0.0.6
httpx==0.25.1

